refactor(console): reorg application details page
