+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Branding/index.module.scss
