+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Branding/index.tsx
