+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Branding/use-application-sign-in-experience-swr.ts
