+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Branding/use-sign-in-experience-swr.ts
