+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Branding/utils.ts
