+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/GuideDrawer/index.module.scss
