+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/MachineLogs/index.module.scss
