+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/MachineLogs/index.tsx
