+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Permissions/index.module.scss
