+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Permissions/index.tsx
