+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Permissions/use-permissions-table.ts
