+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/Settings.tsx
