+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/index.module.scss
