+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/utils.ts
