+++ packages/console/src/pages/ApplicationDetails/GuideModal/index.module.scss
